user_segmentation_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
asyncpg==0.29.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.2
python-multipart==0.0.6 